Add Connect social tab, multi-factor smart workout, injury scanner; real Netanya gyms
Backend (ASP.NET Core):
- Social layer: Friendships, Gyms + GymCoaches, CoachOffers, presence (LastSeen)
  + nearby discovery. New SocialController / GymsController + BL/DAL + SocialModels.
- POCOs relocated from BL/Models to Models (namespace unchanged).

Frontend (Expo):
- Connect tab: map of REAL Netanya gyms (Google Places), nearby people as a list
  (no exact location, for privacy), friend requests, coach offers, filter menu,
  resizable map.
- My Network hub (Coaches/Friends swipe), presence dots + heartbeat, friend chat.
- Multi-factor smart-workout card (collapsible), injury scanner + AI advice.
- SignUp age 18 to 75; Warnings dashboard foldable coach recs + trimmed buttons.

SQL: 2026-06-08_add_social.sql (idempotent) seeds real Netanya gyms + demo users.

+++ TrainWise/TrainWise/Models/ActivityLog.cs
@@ -13,8 +13,8 @@
         public double? CaloriesBurned { get; set; }
         public string SourceDevice { get; set; }
         public byte ExertionLevel { get; set; }        
-        public short Duration { get; set; }               
-        public short CalculatedLoadForSession { get; set; }
+        public int Duration { get; set; }               
+        public int CalculatedLoadForSession { get; set; }
         public bool IsConfirmed { get; set; }
     }
 }
